use systemctl for start/stop if available

+++ lib/commands/command-containerd-stop.bash
@@ -1,6 +1,17 @@
-# steps from https://guide2wsl.com/nerdctl/
-containerDPath=$(asdf which containerd)
-buildkitDPath=$(asdf which buildkitd)
+#!/bin/bash
 
-pkill $containerDPath
-pkill $buildkitDPath
+$containerdStatus=$(systemctl --user is-active containerd)
+if [[ "active" == "$containerdStatus" ]]; then
+    systemctl --user stop containerd
+else
+    containerDPath="${ASDF_INSTALL_PATH}/bin/containerd"
+    pkill $containerDPath
+fi
+
+$buildkitStatus=$(systemctl --user is-active buildkit)
+if [[ "active" == "$buildkitStatus" ]]; then
+    systemctl --user stop buildkit
+else
+    buildkitDPath="${ASDF_INSTALL_PATH}/bin/buildkitd"
+    pkill $buildkitDPath
+fi
